Time: 3 ms (87.54%) | Memory: 8.5 MB (85.52%) - LeetSync

+++ 3768-check-if-digits-are-equal-in-string-after-operations-i/check-if-digits-are-equal-in-string-after-operations-i.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     bool hasSameDigits(std::string s) {
-        int len = s.size();
+        size_t len = s.size();
         while (len > 2) {
-            for (int i = 0; i < len - 1; ++i) {
+            for (size_t i = 0; i < len - 1; ++i) {
                 s[i] = ((s[i] - '0') + (s[i + 1] - '0')) % 10 + '0';
             }
             --len;
